[Foundation] Extract geometry extensions into a seperate file
Non functional change; purely organization/gardening
